install-on-desktops.md: Debian install: use wget and temp file
Instead of a curl we can use the wget that is always installed on all Debian systems.
But since the curl is needed inside of the installation script we still need to install it before. We may get rid of it later.

The wget uses the -nv flag to not show a progress but show any errors.
The script is downloaded first to temp folder and only then being executed. This protects us from executing partially downloaded file.
Instead of piping we executing a bash with the file name and once executed it will be removed.